+++ src/utils/omni-search-util/index.js
@@ -6,5 +6,6 @@ export {
   getOmniTextFromSearchValues,
   getOmniTextFromKeyValues,
   getSearchOptions,
+  getValueItemsFromSearchValues,
   parseValuesFromOmniText,
 } from "./omni-search-util";